Revert "Used "size_t" for all sorting algorithms when indices are stored."
This reverts commit 49e25f4d1be05d3c4649b19e7b2df5510cb5c134.

Using "size_t" causes crashes in quick sort. Reverting to use another
data type.

+++ src/merge_sort.h
@@ -11,10 +11,7 @@ namespace do_sort {
     private:
       vector<int> aux;
 
-      void recursive_msort(vector<int>& v,
-                           vector<int>& aux,
-                           size_t low,
-                           size_t high);
+      void recursive_msort(vector<int>& v, vector<int>& aux, int low, int high);
 
     protected:
       void specific_do_sort();
